+++ changelog.md
@@ -3,8 +3,10 @@
   - Hearts can now optionally be displayed instead of stars for favorites
 - Added ability to search for mods by short description
 - Added optional mod description button in mod list item
+- Added Geode theme compatibility to *View* button in mod list item
 - Set default value for setting *`Only List Enabled Mods`* to `true`
 - Removed toggle setting *`Error Alerts`*
+- Added stats button to pop-up
 - Other tweaks
 
 # v1.0.3